+++ Tools/hvshutdown/hvshutdown.c
@@ -31,10 +31,10 @@ static CFRunLoopSourceRef     sHVShutdownNotificationCFRunLoopSource  = NULL;
 static void hvShutdownNotification(CFMachPortRef port, void *msg, CFIndex size, void *info) {
   HyperVShutdownNotificationMessage *hvMsg = (HyperVShutdownNotificationMessage *) msg;
   HVDBGLOG(stdout, "Received notification of type 0x%X", hvMsg->type);
-  
+
   if (hvMsg->type == kHyperVShutdownNotificationTypePerformShutdown) {
     HVSYSLOG(stdout, "Shutdown request received, performing shutdown");
-    
+
     //
     // Shutdown has been requested, invoke /sbin/shutdown.
     //
@@ -44,13 +44,34 @@ static void hvShutdownNotification(CFMachPortRef port, void *msg, CFIndex size, 
       "now",
       NULL
     };
-    
+
     //
     // This should not return.
     //
     int ret = execv(shutdownArgs[0], shutdownArgs);
     if (ret == -1) {
       HVSYSLOG(stderr, "Failed to execute %s", shutdownArgs[0]);
+    }
+
+  } else if (hvMsg->type == kHyperVShutdownNotificationTypePerformRestart) {
+    HVSYSLOG(stdout, "Restart request received, performing restart");
+
+    //
+    // Restart has been requested, invoke /sbin/shutdown.
+    //
+    char *restartArgs[] = {
+      SHUTDOWN_BIN_PATH,
+      "-r",
+      "now",
+      NULL
+    };
+
+    //
+    // This should not return.
+    //
+    int ret = execv(restartArgs[0], restartArgs);
+    if (ret == -1) {
+      HVSYSLOG(stderr, "Failed to execute %s", restartArgs[0]);
     }
   }
 }
@@ -60,36 +81,36 @@ static void hvShutdownTeardownNotification() {
     IOServiceClose(sConnection);
     sConnection = 0;
   }
-  
+
   if (sDevice != 0) {
     IOObjectRelease(sDevice);
     sDevice = 0;
   }
-  
+
   if (sHVShutdownNotificationCFRunLoopSource != NULL) {
     CFRunLoopRemoveSource(CFRunLoopGetCurrent(), sHVShutdownNotificationCFRunLoopSource, kCFRunLoopCommonModes);
     CFRelease(sHVShutdownNotificationCFRunLoopSource);
     sHVShutdownNotificationCFRunLoopSource = NULL;
   }
-  
+
   if (sHVShutdownNotificationCFMachPort != NULL) {
     CFMachPortInvalidate(sHVShutdownNotificationCFMachPort);
     CFRelease(sHVShutdownNotificationCFMachPort);
     sHVShutdownNotificationCFMachPort = NULL;
   }
-  
+
   if (sHVShutdownNotificationPort != MACH_PORT_NULL) {
     mach_port_destroy(mach_task_self(), sHVShutdownNotificationPort);
     sHVShutdownNotificationPort = MACH_PORT_NULL;
   }
-  
+
   HVDBGLOG(stdout, "Service closed");
 }
 
 static IOReturn hvShutdownSetupNotification(io_service_t device) {
   IOReturn          result;
   CFMachPortContext context;
-  
+
   //
   // Setup notification for shutdown requests.
   //
@@ -98,20 +119,20 @@ static IOReturn hvShutdownSetupNotification(io_service_t device) {
   context.retain = NULL;
   context.release = NULL;
   context.copyDescription = NULL;
-  
+
   result = IOCreateReceivePort(kOSAsyncCompleteMessageID, &sHVShutdownNotificationPort);
   if (result != kIOReturnSuccess) {
     HVSYSLOG(stderr, "Failure while creating notification port: 0x%X", result);
     return result;
   }
-  
+
   sHVShutdownNotificationCFMachPort = CFMachPortCreateWithPort(kCFAllocatorDefault, sHVShutdownNotificationPort, hvShutdownNotification, &context, NULL);
   if (sHVShutdownNotificationCFMachPort == NULL) {
     HVSYSLOG(stderr, "Failed while creating notification CFMachPort");
     hvShutdownTeardownNotification();
     return kIOReturnNoResources;
   }
-  
+
   sHVShutdownNotificationCFRunLoopSource = CFMachPortCreateRunLoopSource(kCFAllocatorDefault, sHVShutdownNotificationCFMachPort, 0);
   if (sHVShutdownNotificationCFRunLoopSource == NULL) {
     HVSYSLOG(stderr, "Failed while creating notification run loop source");
@@ -119,7 +140,7 @@ static IOReturn hvShutdownSetupNotification(io_service_t device) {
     return kIOReturnNoResources;
   }
   CFRunLoopAddSource(CFRunLoopGetCurrent(), sHVShutdownNotificationCFRunLoopSource, kCFRunLoopDefaultMode);
-  
+
   //
   // Connect to service.
   //
@@ -131,7 +152,7 @@ static IOReturn hvShutdownSetupNotification(io_service_t device) {
     return result;
   }
   HVDBGLOG(stdout, "Opened connection to service");
-  
+
   result = IOConnectSetNotificationPort(sConnection, 0, sHVShutdownNotificationPort, 0);
   if (result != kIOReturnSuccess) {
     HVSYSLOG(stderr, "Failed to set notification port: 0x%X", result);
@@ -139,7 +160,7 @@ static IOReturn hvShutdownSetupNotification(io_service_t device) {
     return result;
   }
   HVDBGLOG(stdout, "Port 0x%p setup for notifications", sHVShutdownNotificationPort);
-  
+
   return kIOReturnSuccess;
 }
 
@@ -147,7 +168,7 @@ static void hvShutdownAppeared(void *refCon, io_iterator_t iterator) {
   IOReturn      result;
   io_service_t  obj;
   io_service_t  device = 0;
-  
+
   while ((obj = IOIteratorNext(iterator))) {
     if (!device) {
       device = obj;
@@ -155,7 +176,7 @@ static void hvShutdownAppeared(void *refCon, io_iterator_t iterator) {
       IOObjectRelease(obj);
     }
   }
-  
+
   if (device) {
     result = hvShutdownSetupNotification(device);
     if (result != kIOReturnSuccess) {
@@ -167,7 +188,7 @@ static void hvShutdownAppeared(void *refCon, io_iterator_t iterator) {
 static void hvShutdownRemoved(void *refCon, io_iterator_t iterator) {
   io_service_t  obj;
   bool          deviceIsRemoved = false;
-  
+
   while ((obj = IOIteratorNext(iterator))) {
     if (sDevice && IOObjectIsEqualTo(obj, sDevice)) {
       deviceIsRemoved = true;
@@ -175,7 +196,7 @@ static void hvShutdownRemoved(void *refCon, io_iterator_t iterator) {
 
     IOObjectRelease(obj);
   }
-  
+
   //
   // Close connection to service.
   //
@@ -187,7 +208,7 @@ static void hvShutdownRemoved(void *refCon, io_iterator_t iterator) {
 static IOReturn hvShutdownSetupIOKitNotifications() {
   CFMutableDictionaryRef  matching;
   IOReturn                result;
-  
+
   //
   // Create notification port for I/O Kit notifications.
   //
@@ -196,7 +217,7 @@ static IOReturn hvShutdownSetupIOKitNotifications() {
     HVSYSLOG(stderr, "Failure while creating notification port");
     return kIOReturnNoResources;
   }
-  
+
   //
   // Configure matching notifications.
   // An extra retain call is required as IOServiceAddMatchingNotification
@@ -208,7 +229,7 @@ static IOReturn hvShutdownSetupIOKitNotifications() {
     return kIOReturnNoResources;
   }
   matching = (CFMutableDictionaryRef)CFRetain(matching);
-  
+
   result = IOServiceAddMatchingNotification(sIOKitNotificationPort,
                                             kIOPublishNotification,
                                             matching,
@@ -220,7 +241,7 @@ static IOReturn hvShutdownSetupIOKitNotifications() {
     return result;
   }
   hvShutdownAppeared(NULL, sHVShutdownAppearedIterator);
-  
+
   result = IOServiceAddMatchingNotification(sIOKitNotificationPort,
                                             kIOTerminatedNotification,
                                             matching,
@@ -232,14 +253,14 @@ static IOReturn hvShutdownSetupIOKitNotifications() {
     return result;
   }
   hvShutdownRemoved(NULL, sHVShutdownRemovedIterator);
-  
+
   sIOKitNotificationCFRunLoopSource = IONotificationPortGetRunLoopSource(sIOKitNotificationPort);
   if (sIOKitNotificationCFRunLoopSource == NULL) {
     HVSYSLOG(stderr, "Failed while creating matching notification run loop source");
     return kIOReturnNoResources;
   }
   CFRunLoopAddSource(CFRunLoopGetCurrent(), sIOKitNotificationCFRunLoopSource, kCFRunLoopDefaultMode);
-  
+
   return kIOReturnSuccess;
 }
 
@@ -250,7 +271,7 @@ int main(int argc, const char * argv[]) {
   if (hvShutdownSetupIOKitNotifications() != kIOReturnSuccess) {
     return -1;
   }
-  
+
   //
   // Run main loop, this should not return.
   //